Refactor/bill edit form (#117)
* refactor: ui desgins

* fix: removed unnecessary file

* refactor(bill edit form): decoupled compoenents

* refactor: changed bill form ui
